fix: Address code-quality bot review findings
- Merge the empty UnauthorizedAccessException catch into the
  IOException one in RemoteControlClient.IsAndroidEmulator, so
  no catch block is empty.
- Cast to double before squaring in PanFromCenter's distance
  calculation to avoid integer-overflow risk.
- Make the null check explicit for targetFramework in
  XamlCodeGeneration.GetUnoRuntime so static analysis can prove
  the Substring call safe.

+++ src/Uno.UI.RuntimeTests/IntegrationTests/common/TestServices.InputHelper.cs
@@ -131,7 +131,7 @@ namespace Private.Infrastructure
 					var center = new Point(topLeft.X + element.RenderSize.Width / 2, topLeft.Y + element.RenderSize.Height / 2);
 
 					const uint steps = 10;
-					var durationInMilliseconds = Math.Sqrt(relX * relX + relY * relY) / velocityFactor;
+					var durationInMilliseconds = Math.Sqrt((double)relX * relX + (double)relY * relY) / velocityFactor;
 
 					finger.Press(center);
 					finger.MoveTo(new Point(center.X + relX, center.Y + relY), steps, (uint)Math.Max(1, durationInMilliseconds / steps));
